move schema-gen to common
