+++ common/database/src/testFixtures/kotlin/no/nav/mulighetsrommet/database/kotest/extensions/DatabaseTestConfig.kt
@@ -1,16 +1,16 @@
-package no.nav.mulighetsrommet.arena.adapter.no.nav.mulighetsrommet.arena.adapter.utils
+package no.nav.mulighetsrommet.database.kotest.extensions
 
 import no.nav.mulighetsrommet.database.DatabaseConfig
 import no.nav.mulighetsrommet.database.Password
 import java.util.*
 
-fun createDatabaseConfigWithRandomSchema(
+fun createDatabaseTestSchema(
+    name: String,
+    port: Int,
     host: String = "localhost",
-    port: Int = 5443,
-    name: String = "mulighetsrommet-arena-adapter-db",
     user: String = "valp",
     password: Password = Password("valp")
 ): DatabaseConfig {
-    val schema = "${UUID.randomUUID()}"
+    val schema = "$port--${UUID.randomUUID()}"
     return DatabaseConfig(host, port, name, schema, user, password, 1)
 }
