+++ mulighetsrommet-arena-adapter/src/test/kotlin/no/nav/mulighetsrommet/arena/adapter/kafka/KafkaConsumerOrchestratorTest.kt
@@ -12,10 +12,10 @@ import io.mockk.mockk
 import io.mockk.verify
 import kotlinx.coroutines.delay
 import no.nav.common.kafka.util.KafkaPropertiesBuilder
-import no.nav.mulighetsrommet.arena.adapter.no.nav.mulighetsrommet.arena.adapter.utils.createDatabaseConfigWithRandomSchema
 import no.nav.mulighetsrommet.arena.adapter.repositories.Topic
 import no.nav.mulighetsrommet.arena.adapter.repositories.TopicRepository
 import no.nav.mulighetsrommet.database.kotest.extensions.FlywayDatabaseListener
+import no.nav.mulighetsrommet.database.kotest.extensions.createDatabaseTestSchema
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.common.serialization.ByteArrayDeserializer
 import org.testcontainers.containers.KafkaContainer
@@ -28,7 +28,7 @@ internal class KafkaConsumerOrchestratorTest : FunSpec({
 
     lateinit var consumerProperties: Properties
     val listener =
-        FlywayDatabaseListener(createDatabaseConfigWithRandomSchema())
+        FlywayDatabaseListener(createDatabaseTestSchema("mulighetsrommet-arena-adapter-db", 5443))
     register(listener)
 
     val kafka = install(
